+++ src/App.jsx
@@ -5,6 +5,7 @@ import { Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
 import Shop from "./pages/Shop";
 import Cart from "./pages/Cart";
+import Category from "./pages/[category]";
 
 function App() {
   const [searchQuery, setSearchQuery] = useState("");
@@ -23,6 +24,7 @@ function App() {
             <Route path="/" element={<Home />} />
             <Route path="/shop" element={<Shop />} />
             <Route path="/cart" element={<Cart />} />
+            <Route path="/category/:category" element={<Category />} />
           </Routes>
         </div>
