+++ src/components/ProductList2.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function ProductList2() {
+  return (
+    <div>ProductList2</div>
+  )
+}
+
+export default ProductList2
