+++ src/pages/[category].jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export default function category() {
+
+  //bruk fetch til å hente produktene i kategorien
+  // dynamic routes, https://nextjs.org/docs/routing/dynamic-routes / display enten ID eller slug i URL.
+  return (
+    < div className="category-parent" >
+      <h1>Hello world</h1>
+    </ div >
+  );
+}
